refactor(preprints-models): Decomposed, used shared models. Moved some to shared
