+++ src/app/features/preprints/constants/index.ts
@@ -1,3 +1,2 @@
 export * from './form-input-limits.const';
-export * from './preprints.routes';
 export * from './submit-preprint-steps.const';
