+++ src/app/features/preprints/models/index.ts
@@ -1,3 +1,6 @@
+export * from './preprint.models';
+export * from './preprint-json-api.models';
 export * from './preprint-licenses-json-api.models';
-export * from './preprints.models';
+export * from './preprint-provider.models';
+export * from './preprint-provider-json-api.models';
 export * from './submit-preprint-form.models';
