+++ src/app/features/preprints/store/preprint-providers/preprint-providers.model.ts
@@ -1,5 +1,5 @@
-import { PreprintProviderDetails, PreprintProviderShortInfo, Subject } from '@osf/features/preprints/models';
-import { AsyncStateModel } from '@shared/models';
+import { PreprintProviderDetails, PreprintProviderShortInfo } from '@osf/features/preprints/models';
+import { AsyncStateModel, Subject } from '@shared/models';
 
 export interface PreprintProvidersStateModel {
   preprintProvidersDetails: AsyncStateModel<PreprintProviderDetails[]>;
