+++ src/app/shared/models/subject/subjects-json-api.model.ts
@@ -10,12 +10,16 @@ export type SubjectDataJsonApi = ApiData<
   SubjectAttributesJsonApi,
   SubjectEmbedsJsonApi,
   SubjectRelationshipsJsonApi,
-  null
+  SubjectLinksJsonApi
 >;
 
 interface SubjectAttributesJsonApi {
   text: string;
   taxonomy_name: string;
+}
+
+interface SubjectLinksJsonApi {
+  iri: string;
 }
 
 interface SubjectRelationshipsJsonApi {
